Lexer
The latest development of the lexer has resulted in notable advancements, specifically in the areas of string handling, operator recognition, boolean identification, keyword and identifier recognition, as well as support for integers, floats, and multi-operators (such as ">>" and ">="). With these updates, a new list of tokens is generated for each line of code within the language, which is then compiled into a nested list for comprehensive parsing. As a result, the lexer is now capable of fully processing code files and producing a complete and accurate nested list of tokens for further analysis and manipulation. These improvements represent significant strides towards enhancing the functionality and versatility of the lexer, and have established a more robust foundation for continued development and refinement in the future.