Implemented Null. Range, debugged assignOp, updateOp, print
Implemented class Range in data type declaration which helps “for” loop to get a list from integer “num1” to integer “num2”with “jmp” jumps.
Implemented “Range” in evalAST.py.
Debugged assignOp and updateOp: Previously assignOp and updateOp were first evaluating the value and then storing in the variable but now it directly stored the AST of value into the variable for consistency.
Debugged listOp: Now it can handle lists as List_ and Variable both. Now we can pass a list in the form of [1,2,3] or a variable containing that list.
Debugged print: If the data type of element that we need to print is variable, we need to double eval it.